remove 'error' status from being returned
I think we have an issue on how errors are being categorized as ‘error’
vs ‘failure’.
Conduit core is expecting ‘error’ status to be  ‘connection_timeout’
(‘timeout’, ‘error’) status. But the conduit_reactor driver is looking
for no ‘!response_content’ (response content at the moment is presence
of the resource in the json response). The problem here is that reactor
can return errors, errors will not include the obj resource like the
failure test are including. We should have a conversation/pair session
to determine the correct status ‘error’ vs ‘failure’.